# Faster cumsum and cumprod backwards